Fixed Youtube issue
Once the user authorizes to play, autoplay is enabled in the future on
that particular URL. window.location.href tracking is added to prevent
autoplay on dynamic websites like Youtube, where the video is changed,
but the page is not reloaded. In that case, Youtube changes the URL via
HTML5 history API.

The site is free to autoplay advertisements on the same video or
article URL. This is appropriate, since an Ad-Blocker is needed to skip
these advertisements, and since they are generally continuous with the
video playing.

The main part of the fix has been tested for months and definitely
fixes the issue with Youtube videos.