Import change 67a2b7c1d2fc2d9cd31f3675e59e7a02a8dabc22 from pull request #64
Description from original commit:

Once the user authorizes to play, autoplay is enabled in the future on
that particular URL. window.location.href tracking is added to prevent
autoplay on dynamic websites like Youtube, where the video is changed,
but the page is not reloaded. In that case, Youtube changes the URL via
HTML5 history API.

The site is free to autoplay advertisements on the same video or
article URL. This is appropriate, since an Ad-Blocker is needed to skip
these advertisements, and since they are generally continuous with the
video playing.